Rework the ACK system:
Because ACKs are cumulative there's actually no reason to queue ACK numbers. Just send the latest/most up to date cur_ack number.
Furthermore, each loop in normal_loop actually only receives one packet at once and also immediately exhausts all queued ACKs,
this defeats the point of having any queue or even a count of ACKs, since it's at most 1.

Thus we just have a boolean for if we received a payload packet and thus need to ACK it (or send a dup ack if out of order).

Also, if we receive ACKs that are under the sender window SEQ, if it's not matching last_dup_ack check if it's less than last_dup_ack because if
it is, we don't need to change last_dup_ack since this is clearly an out of order ACK (as we already have received a higher ACK).